moved game to own module.
