+++ src/ui.cc
@@ -1,10 +1,3 @@
-extern "C" {
-#ifndef _WIN32
-#include <ncurses.h>
-#else
-#include <curses.h>
-#endif
-}
 #include "ui.hh"
 #include <cstdarg>
